+++ TowerDefense/Tower.cs
@@ -15,7 +15,7 @@ namespace TowerDefense
 
         public Tower(Texture2D tex, Vector2 position) : base(tex, position) { }
 
-        public abstract List<Projectile> Attack(List<Enemy> enemylist, List<Projectile> projectilelist, int elapsedTime);
+        public abstract List<Projectile> Attack(List<Enemy> enemylist, List<Projectile> projectilelist, double elapsedTime);
 
         public abstract void ShowStats(SpriteBatch batch, SpriteFont font, Viewport viewport);
